+++ src/securejoin.rs
@@ -422,7 +422,7 @@ async fn send_handshake_msg(
     Ok(())
 }
 
-async fn chat_id_2_contact_id(context: &Context, contact_chat_id: ChatId) -> u32 {
+async fn chat_id_2_contact_id(context: &Context, contact_chat_id: ChatId) -> i64 {
     if let [contact_id] = chat::get_chat_contacts(context, contact_chat_id).await[..] {
         contact_id
     } else {
@@ -469,7 +469,7 @@ pub(crate) enum HandshakeError {
     NotSecureJoinMsg,
     #[error("Failed to look up or create chat for contact #{contact_id}")]
     NoChat {
-        contact_id: u32,
+        contact_id: i64,
         #[source]
         cause: Error,
     },
@@ -511,7 +511,7 @@ pub(crate) enum HandshakeMessage {
 pub(crate) async fn handle_securejoin_handshake(
     context: &Context,
     mime_message: &MimeMessage,
-    contact_id: u32,
+    contact_id: i64,
 ) -> Result<HandshakeMessage, HandshakeError> {
     if contact_id <= DC_CONTACT_ID_LAST_SPECIAL {
         return Err(HandshakeError::SpecialContactId);
@@ -939,7 +939,7 @@ pub(crate) async fn handle_securejoin_handshake(
 pub(crate) async fn observe_securejoin_on_other_device(
     context: &Context,
     mime_message: &MimeMessage,
-    contact_id: u32,
+    contact_id: i64,
 ) -> Result<HandshakeMessage, HandshakeError> {
     if contact_id <= DC_CONTACT_ID_LAST_SPECIAL {
         return Err(HandshakeError::SpecialContactId);
@@ -1016,7 +1016,7 @@ pub(crate) async fn observe_securejoin_on_other_device(
 }
 
 async fn secure_connection_established(context: &Context, contact_chat_id: ChatId) {
-    let contact_id: u32 = chat_id_2_contact_id(context, contact_chat_id).await;
+    let contact_id: i64 = chat_id_2_contact_id(context, contact_chat_id).await;
     let contact = Contact::get_by_id(context, contact_id).await;
 
     let addr = if let Ok(ref contact) = contact {
